requirements.txt
certifi==2020.6.20
elasticsearch==7.9.1
python-dateutil==2.8.1
six==1.15.0
urllib3==1.25.11
